refactor: split auth upgrade prompt sheets
